initial restructuring to modulith
