+++ src/main/java/ukma/edu/ua/HospitalApp/doctor/DoctorDTO.java
