+++ src/main/java/ukma/edu/ua/HospitalApp/patient/PatientMapper.java
@@ -5,7 +5,7 @@ import org.mapstruct.Mapping;
 import org.mapstruct.MappingConstants;
 import org.mapstruct.factory.Mappers;
 import ukma.edu.ua.HospitalApp.dto.PatientDTO;
-import ukma.edu.ua.HospitalApp.models.PatientDetails;
+import ukma.edu.ua.HospitalApp.patient.PatientDetails;
 
 @Mapper(componentModel = MappingConstants.ComponentModel.SPRING)
 public interface PatientMapper {
