+++ src/main/java/ukma/edu/ua/HospitalApp/prescription/Drug.java
@@ -9,6 +9,7 @@ import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
 import lombok.experimental.SuperBuilder;
+import ukma.edu.ua.HospitalApp.prescription.Prescription;
 
 @EqualsAndHashCode(callSuper = true)
 @Data
